backend/requirements.txt
numpy
pillow
flask
flask-cors
tensorflow
opencv-python
scikit-image

# Database & Auth
Flask-MongoEngine
Flask-JWT-Extended
Authlib
python-dotenv
google-auth
